+++ Code/Components/Services/skymodel/current/schema/DataSource.h
@@ -63,7 +63,7 @@ struct DataSource {
     // @brief Optimistic concurrency lock version
     // @units none
     #pragma db version
-    boost::uint32_t version;
+    boost::uint64_t version;
 
     // @brief Primary key unique identifier
     // @units none
